refactor(Director): Move google creds to new file
Need to allow the google credentials code to be imported independently
of the google docs facade.
